requirements.txt
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
SQLAlchemy==1.4.46
pytest
parameterized